requirements.txt
fastapi
uvicorn
tensorflow-cpu==2.15.0
pillow
numpy
python-multipart